+++ CHANGELOG.md
@@ -6,5 +6,6 @@ All dates listed follow the ISO standard of yyyy-mm-dd.
 
 ## [Unreleased]
 ### Added
+ - median_of_squares function to compute median of squared residuals [David Richardson <drichardson42@gatech.edu>]
 ### Changed
 ### Removed